add admin power
add admin power

+++ app/controller/AdminController.php
@@ -10,6 +10,12 @@ use HFB\app\model\db\UserModel;
 
 class AdminController extends Controller {
 	
+	public function redirect() {
+		if(!isset($_SESSION['admin'])) {
+			return self::_alertRedirect("请先登录!", "/admin/login");
+		}
+	}
+	
 	public function login() {
 		
 	}
@@ -50,19 +56,17 @@ class AdminController extends Controller {
 	}
 	
 	public function houseInfo($page = 1) {
+		$this->redirect();
+	
 		$_SESSION['link'] = "houseInfo";
 		
-		static $pageSize = 4;
+		static $pageSize = 8;
 		static $orderby = "id desc";
 		$data['isSelled'] = "0";
 		
-		if(!isset($_SESSION['admin'])) {
-			return self::_alertRedirect("请先登录!", "/admin/login");
-		}
-	
 		$houseInfoModel = new HouseinfoModel();
 		$houses = $houseInfoModel->getListByMap($data);
-	
+		
 		$count = count($houses);
 		$pages = ceil($count / $pageSize);
 		
@@ -70,18 +74,18 @@ class AdminController extends Controller {
 		$countHouseInfo = count($housesInfo);
 		
 		$publishers = array();
-	
+		
 		for($i=0; $i<$countHouseInfo; $i++) {
 			$map['id'] = $housesInfo[$i]['sellerId'];
-				
+		
 			$userModel = new UserModel();
 			$res = $userModel->getByMap($map);
-				
+		
 			$username = $res['username'];
-				
+		
 			array_push($publishers, $username);
 		}
-	
+		
 		self::_bindValue("username", $_SESSION['admin']);
 		self::_bindValue("houses", $housesInfo);
 		self::_bindValue("publishers", $publishers);
@@ -89,37 +93,51 @@ class AdminController extends Controller {
 	}
 	
 	public function deleteHouseInfo() {
-		$map['id'] = $_GET['id'];
+		$power = "房源管理权限";
+		$check = $this->checkPower($power);
 		
-		$houseInfoModel = new HouseinfoModel();
-		$res = $houseInfoModel->delete($map);
-		
-		if($res) {
-			return self::_alertRedirect("删除成功!");
+		if($check) {
+			$map['id'] = $_GET['id'];
+			
+			$houseInfoModel = new HouseinfoModel();
+			$res = $houseInfoModel->delete($map);
+			
+			if($res) {
+				return self::_alertRedirect("删除成功!");
+			}else {
+				return self::_alertRedirect("删除失败!");
+			}
 		}else {
-			return self::_alertRedirect("删除失败!");
+			return self::_alertRedirect("您没有".$power."权限!");
 		}
 	}
 	
 	public function checkHouseInfo() {
-		$check = $_GET['check'];
-		$map['id'] = $_GET['id'];
+		$power = "房源管理权限";
+		$check = $this->checkPower($power);
 		
-		if("true" == $check) {
-			$data['checkPass'] = "1";
+		if($check) {
+			$check = $_GET['check'];
+			$map['id'] = $_GET['id'];
+			
+			if("true" == $check) {
+				$data['checkPass'] = "1";
+			}else {
+				$data['checkPass'] = "2";
+			}
+			$houseInfoModel = new HouseinfoModel();
+			
+			$res = $houseInfoModel->update($data, $map);
+			
+			if($res) {
+				return self::_alertRedirect("审核成功!");
+			}else {
+				return self::_alertRedirect("审核失败!");
+			}
 		}else {
-			$data['checkPass'] = "2";
+			return self::_alertRedirect("您没有".$power."权限!");
 		}
 		
-		$houseInfoModel = new HouseinfoModel();
-		
-		$res = $houseInfoModel->update($data, $map);
-		
-		if($res) {
-			return self::_alertRedirect("审核成功!");
-		}else {
-			return self::_alertRedirect("审核失败!");
-		}
 	}
 	
 	public function houseDetail() {
@@ -136,14 +154,12 @@ class AdminController extends Controller {
 	}
 	
 	public function sellRecord($page = 1) {
+		$this->redirect();
+		
 		$_SESSION['link'] = "sellRecord";
-		static $pageSize = 4;
+		static $pageSize = 8;
 		static $orderby = "id desc";
 		$data['isSelled'] = "1";
-		
-		if(!isset($_SESSION['admin'])) {
-			return self::_alertRedirect("请先登录!", "/admin/login");
-		}
 		
 		$houseInfoModel = new HouseinfoModel();
 		$houses = $houseInfoModel->getListByMap($data);
@@ -171,5 +187,155 @@ class AdminController extends Controller {
 		self::_bindValue("houses", $housesInfo);
 		self::_bindValue("publishers", $publishers);
 		self::_bindValue("pages", $pages);
+	}
+	
+	public function userInfo($page = 1) {
+		$this->redirect();
+		
+		$_SESSION['link'] = "userInfo";
+		
+		static $pageSize = 8;
+		static $orderby = "id desc";
+		$data = array();
+		
+		if(!isset($_SESSION['admin'])) {
+			return self::_alertRedirect("请先登录!", "/admin/login");
+		}
+		
+		$userModel = new UserModel();
+		$users = $userModel->getListByMap($data);
+		
+		$count = count($users);
+		$pages = ceil($count / $pageSize);
+		
+		$usersInfo = $userModel->getListByMap($data, $page, $pageSize, $orderby);
+		$countUserInfo = count($usersInfo);
+		
+		$publishers = array();
+
+		self::_bindValue("username", $_SESSION['admin']);
+		self::_bindValue("users", $usersInfo);
+		self::_bindValue("pages", $pages);
+	}
+	
+	public function deleteUserInfo() {
+		$power = "用户管理权限";
+		$check = $this->checkPower($power);
+		
+		if($check) {
+			$map['id'] = $_GET['id'];
+			
+			$userModel = new UserModel();
+			$res = $userModel->delete($map);
+			
+			if($res) {
+				return self::_alertRedirect("删除成功!");
+			}else {
+				return self::_alertRedirect("删除失败!");
+			}
+		}else {
+			return self::_alertRedirect("您没有".$power."权限!");
+		}
+	}
+	
+	public function addAdmin() {
+		$power = "管理员管理权限";
+		$check = $this->checkPower($power);
+		
+		if($check) {
+			$this->redirect();
+		
+			$_SESSION['link'] = "adminInfo";
+		}else {
+			return self::_alertRedirect("您没有".$power."权限!");
+		}
+	}
+	
+	public function doAddAdmin() {
+		$username = $_POST['username'];
+		$password = $_POST['password'];
+		$email = $_POST['email'];
+		
+		if(isset($_REQUEST['power'])) {
+			$power = "";
+			foreach($_POST['power'] as $checkbox) {
+				$power .= $checkbox.";";
+			}
+		}else {
+			$power = "";
+		}
+		
+		$data['username'] = $username;
+		$data['password'] = md5($password);
+		$data['email'] = $email;
+		$data['power'] = $power;
+		
+		$check = $this->checkAddAdmin($data, $password);
+		
+		if($check) {
+			$adminModel = new AdminModel();
+			$res = $adminModel->add($data);
+			
+			if($res) {
+				return self::_genJSONResult(['code' => 0, 'msg' => '添加成功!', 'redirect' => '/admin/updateAdmin']);
+			}else {
+				return self::_genJSONResult(['code' => -1, 'msg' => '添加失败!']);
+			}
+		}
+		
+	}
+	
+	public function updateAdmin($data) {
+		$this->redirect();
+		
+		$_SESSION['link'] = "adminInfo";
+	}
+	
+	public function checkAddAdmin($data, $password) {
+		self::_setExceptionHandler('HFB\app\exception\JSONExceptionHandler');
+		
+		if("" == $data['username']) {
+			throw new Exception("用户名不能为空!");
+		}else {
+			$adminModel = new AdminModel();
+			$map['username'] = $data['username'];
+			$res = $adminModel->getByMap($map);
+			
+			if($res) {
+				throw new Exception("该用户名已被注册!");
+			}
+		}
+		if("" == $password) {
+			throw new Exception("密码不能为空!");
+		}
+		if("" == $data['email']) {
+			throw new Exception("邮箱不能为空!");
+		}else {
+			$adminModel = new AdminModel();
+			$data2['email'] = $data['email'];
+			$res = $adminModel->getByMap($data2);
+			 
+			if($res) {
+				throw new Exception("该邮箱已被注册!");
+			}
+		}
+		
+		return TRUE;
+	}
+	
+	private function checkPower($power) {
+		$flag = false;
+		$map['username'] = $_SESSION['admin'];
+		
+		$adminModel = new AdminModel();
+		$res = $adminModel->getByMap($map);
+		$powers = explode(";", $res['power']);
+		
+		foreach ($powers as $_power) {
+			if($power == $_power) {
+				$flag = true;
+			}
+		}
+		return $flag;
 	}
 }
